+++ house/wall_tab.go
@@ -91,8 +91,8 @@ func (w *WallPanel) Respond(ui *gui.Gui, group gui.EventGroup) bool {
       if w.viewer.Temp.WallTexture != nil {
         wx,wy := w.viewer.BoardToWindow(w.viewer.Temp.WallTexture.X, w.viewer.Temp.WallTexture.Y)
         px,py := event.Key.Cursor().Point()
-        w.drag_anchor.X = float32(px) - wx - 0.5
-        w.drag_anchor.Y = float32(py) - wy - 0.5
+        w.drag_anchor.X = float32(px) - float32(wx) - 0.5
+        w.drag_anchor.Y = float32(py) - float32(wy) - 0.5
       }
     }
     return true
